@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { App } from './components';
+import { BrowserRouter } from 'react-router-dom';
+import { createRoot } from "react-dom/client"
+import  App  from './components/App';
 // css stylesheets can be created for each component
 // place them in the src/style directory, and import them like this:
 import './style/index.css';
-
-ReactDOM.render(<App />, document.getElementById('root'));
+const root = createRoot(document.querySelector("#root"))
+root.render(<BrowserRouter><App /></BrowserRouter>);
